+++ workspace/lint-staged.config.js
@@ -5,12 +5,14 @@
 const config = {
     // REF: https://github.com/sudo-suhas/lint-staged-django-react-demo
     "*": "cspell lint --no-must-find-files",
-    "*.{js,ts,md,json,yaml}": "eslint --fix",
+    "*.{js,ts,md,json,toml,yaml}": "eslint --fix",
     "*.go": "gofmt -w",
     "*.java": "clang-format -i --style=file:./.clang-format",
-    // "*.java": (_) => "./solutions/java/gradlew -p ./solutions/java spotlessApply",
+    // "*.java": (_) => "./solutions/java/gradlew -p ./solutions/java spotlessJavaApply",
+    "*.gradle": (_) => "./solutions/java/gradlew -p ./solutions/java spotlessGroovyGradleApply",
     // "*.{kt,kts}": "ktlint -F",
-    "*.{kt,kts}": (_) => "./solutions/kotlin/gradlew -p ./solutions/kotlin spotlessApply",
+    "*.{kt,kts}": (_) => "./solutions/kotlin/gradlew -p ./solutions/kotlin spotlessKotlinApply",
+    "*.py": "rye run --pyproject ./solutions/python/pyproject.toml ruff format",
 };
 
 export default config;