+++ workspace/solutions/go/s0/s0/s0/s1/s_0001_test.go
@@ -1,0 +1,18 @@
+package _0001
+
+import (
+	"testing"
+
+	"github.com/Zuoqiu-Yingyi/leetcode-multilingual/workspace/solutions/go/utils"
+
+	"github.com/Zuoqiu-Yingyi/leetcode-multilingual/workspace/solutions/go/s0/s0/s0/s1/s_0001_00"
+)
+
+func TestSolutions(t *testing.T) {
+	utils.Test(
+		[]any{
+			s_0001_00.Solution,
+		},
+		t,
+	)
+}