+++ workspace/solutions/java/lib/src/test/java/s0/s0/s0/s1/s_0001_test.java
@@ -1,0 +1,17 @@
+package s0.s0.s0.s1;
+
+import org.junit.jupiter.api.Test;
+import s0.s0.s0.s1.s_0001_00.s_0001_00;
+import utils.Test_;
+
+class s_0001_test {
+    @Test
+    void test()
+    {
+        Test_ test = new Test_(
+            this.getClass(),
+            s_0001_00.class
+        );
+        test.run();
+    }
+}